Moved view manager and switch event to views package
