+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/FrontendApplication.java
@@ -1,5 +1,6 @@
 package dev.totallyspies.spydle.frontend;
 
+import dev.totallyspies.spydle.frontend.views.ViewManagerModel;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
