+++ frontend/src/main/java/dev/totallyspies/spydle/frontend/interface_adapters/game_room/GameRoomViewController.java
@@ -1,6 +1,6 @@
 package dev.totallyspies.spydle.frontend.interface_adapters.game_room;
 
-import dev.totallyspies.spydle.frontend.SwitchViewEvent;
+import dev.totallyspies.spydle.frontend.views.SwitchViewEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
